text-moderation/pyproject.toml
[project]
name = "text-moderation"
version = "0.1.0"
description = "Text Moderation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "uvicorn",
    "fastapi",
    "openai",
    "pydantic",
    "python-dotenv>=1.1.0",
]
